fix: scoreboard making several requests to the Cartesi Node
